backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
celery==5.3.6
redis==4.6.0
httpx==0.26.0
pillow==10.2.0
pydantic==2.5.3
python-dotenv==1.0.1
python-multipart==0.0.6
pgvector==0.2.4
tenacity==8.2.3
structlog==25.4.0
fastapi-cache2[redis]==0.2.1
slowapi==0.1.9
python-magic==0.4.27